enhancements
Added Conventions for CSS and JS file locations.
Plugin is now using jquery-ui that is shipped with vanilla.
Removed permission requirements
Scripts are not added if user is signed in
Moved all html into a view
Window scrolls to top when sign in is clicked
Username field is focused after clicking sign in